FEATURES: model arg, interpreter env, local .env
model arg: I made the 3rd argument take the model name string so gpt-3.5-turbo could be used.
interpreter env: added the comment to the top of index.js to ensure the node environment.
local .env: I added the path option to dotenv.config() so the symlink can be used in other projects.